GeofenceViewModelとLocationViewModelをMapViewModelに統合
