+++ app/src/main/java/com/example/fitbattleandroid/MainActivity.kt
@@ -23,7 +23,7 @@ import androidx.lifecycle.lifecycleScope
 import com.example.fitbattleandroid.ui.navigation.App
 import com.example.fitbattleandroid.ui.screen.isBackgroundLocationPermissionGranted
 import com.example.fitbattleandroid.ui.theme.FitBattleAndroidTheme
-import com.example.fitbattleandroid.viewmodel.LocationViewModel
+import com.example.fitbattleandroid.viewmodel.MapViewModel
 import com.google.android.gms.common.api.ResolvableApiException
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.location.LocationSettingsRequest
@@ -55,7 +55,7 @@ class MainActivity : ComponentActivity() {
             }
         }
 
-    private val locationViewModel: LocationViewModel by viewModels()
+    private val mapViewModel: MapViewModel by viewModels()
     private val backgroundPermissionGranted = mutableStateOf(false)
 
     @RequiresApi(Build.VERSION_CODES.Q)
@@ -88,7 +88,7 @@ class MainActivity : ComponentActivity() {
         }
 
         // 位置情報の設定を確認
-        checkLocationSettings(locationViewModel)
+        checkLocationSettings(mapViewModel)
 
         val requestPermissionLauncher =
             registerForActivityResult(
@@ -115,7 +115,7 @@ class MainActivity : ComponentActivity() {
                         Modifier
                             .fillMaxSize(),
                     requestPermissionLauncher = requestPermissionLauncher,
-                    locationViewModel = locationViewModel,
+                    mapViewModel = mapViewModel,
                     backgroundPermissionGranted = backgroundPermissionGranted,
                     healthConnectClient = healthConnectClient,
                 )
@@ -151,11 +151,11 @@ class MainActivity : ComponentActivity() {
          */
     }
 
-    private fun checkLocationSettings(locationViewModel: LocationViewModel) {
+    private fun checkLocationSettings(mapViewModel: MapViewModel) {
         val builder =
             LocationSettingsRequest
                 .Builder()
-                .addLocationRequest(locationViewModel.locationRequest)
+                .addLocationRequest(mapViewModel.locationRequest)
         val client: SettingsClient = LocationServices.getSettingsClient(this)
 
         val task: Task<LocationSettingsResponse> = client.checkLocationSettings(builder.build())
